requirements-ocr.txt
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0
